fix: shift related overlay code
