+++ src/AppContainer.js
@@ -1,12 +1,10 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import { Route, Switch } from "react-router-dom";
 import { hot } from "react-hot-loader/root";
-import { CSSTransition } from "react-transition-group";
 import { routes } from "./routes";
 import { NavigationBar } from "./components/Layout/NavigationBar";
 import { FooterDefault } from "./components/Layout/Footer";
 import { Overlay } from "./components/UI/Overlay";
-import { useKeyPress } from "./common/hooks/useKeyPress";
 import { OverlayProvider } from "./common/contexts/OverlayContext";
 
 const AppContainer = () => {
@@ -18,23 +16,6 @@ const AppContainer = () => {
     setOverlayContent,
     isOverlayVisible,
     setOverlayVisible,
-  };
-  const escapePress = useKeyPress("Escape");
-
-  useEffect(() => {
-    if (escapePress) {
-      setOverlayVisible(false);
-    }
-  }, [escapePress]);
-
-  useEffect(() => {
-    if (overlayContent !== null) {
-      setOverlayVisible(true);
-    }
-  }, [overlayContent]); // callback when setOverlayContent is called
-
-  const onOverlayTransitionEnded = () => {
-    setOverlayContent(null);
   };
 
   return (
@@ -49,15 +30,7 @@ const AppContainer = () => {
           </Switch>
         </main>
         <FooterDefault />
-        <CSSTransition
-          in={isOverlayVisible}
-          timeout={400}
-          classNames="fade"
-          unmountOnExit
-          onExited={onOverlayTransitionEnded}
-        >
-          <Overlay>{overlayContent}</Overlay>
-        </CSSTransition>
+        <Overlay>{overlayContent}</Overlay>
       </OverlayProvider>
     </>
   );
